+++ packages/webblocks-cms/src/Support/Plugins/Catalog/PluginCatalogResult.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace WebBlocks\Cms\Support\Plugins\Catalog;
+
+class PluginCatalogResult
+{
+  /**
+   * @param  array<int, CatalogPlugin>  $plugins
+   */
+  public function __construct(
+    public readonly bool $available,
+    public readonly array $plugins,
+    public readonly string $baseUrl,
+    public readonly string $cmsVersion,
+    public readonly ?string $message = null,
+  ) {}
+}